+++ src/app/dashboard/dashboard.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
+import { Hero } from '../shared/models/hero';
+import { HeroService } from '../shared/services/hero.service';
 
 @Component({
   selector: 'app-dashboard',
