+++ src/app/hero-search/hero-search.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable, Subject } from 'rxjs';
 import { debounceTime, distinctUntilChanged, switchMap } from 'rxjs/operators';
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
+import { Hero } from '../shared/models/hero';
+import { HeroService } from '../shared/services/hero.service';
 
 @Component({
   selector: 'app-hero-search',
