+++ src/app/heroes/hero-list/hero-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
+import { Hero } from 'src/app/shared/models/hero';
+import { HeroService } from 'src/app/shared/services/hero.service';
 
 @Component({
   selector: 'app-heroes',
