+++ src/app/in-memory-data.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { InMemoryDataService } from './in-memory-data.service';
+import { InMemoryDataService } from './shared/services/in-memory-data.service';
 
 describe('InMemoryDataService', () => {
   let service: InMemoryDataService;
