+++ src/app/message.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { MessageService } from './message.service';
+import { MessageService } from './shared/services/message.service';
 
 describe('MessageService', () => {
   let service: MessageService;
