+++ src/app/mock-heroes.ts
@@ -1,4 +1,4 @@
-import { Hero } from "./hero";
+import { Hero } from "./shared/models/hero";
 
 export const HEROES: Hero[] = [
   { id: 11, name: 'Dr Nice' },
