+++ src/app/shared/services/hero.service.ts
@@ -1,9 +1,9 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
-import { Hero } from './hero';
+import { Hero } from '../models/hero';
 import { MessageService } from './message.service';
-import { HEROES } from './mock-heroes';
+import { HEROES } from '../../mock-heroes';
 import { catchError, map, tap } from 'rxjs/operators';
 
 @Injectable({
